requirements.txt
streamlit
requests
pandas
python-dotenv
sqlalchemy>=2.0.0
tqdm
